+++ speeduino/src/decoders/missing_tooth.cpp
@@ -497,8 +497,8 @@ int getCrankAngle_missingTooth(void)
     crankAngle += 360;
   }
 
-  unsigned long const lastCrankAngleCalc = micros();
-  unsigned long const elapsedTime = lastCrankAngleCalc - tempToothLastToothTime;
+  uint32_t const lastCrankAngleCalc = micros();
+  uint32_t const elapsedTime = lastCrankAngleCalc - tempToothLastToothTime;
   crankAngle += timeToAngleDegPerMicroSec(elapsedTime, degreesPerMicro);
 
   if (crankAngle >= 720)
